// server/openstorefront/openstorefront-web/src/main/ngapp/src/app/Models/Branding.ts
export class Branding {
    brandingId: string;
    name: string;
    primaryLogoUrl: string;
    secondaryLogoUrl: string;
    homebackSplashUrl: string;
    loginFooter: string;
    loginContentBlock: string;
    loginLogoUrl: string;
    loginOverviewVideoUrl: string;
    loginOverviewVideoPosterUrl: string;
    loginRegistrationVideoUrl: string;
    applicationName: string;
    landingPageTitle: string;
    landingStatsText: string;
    landingPageBanner: string;
    landingPageFooter: string;
    analyticsTrackingCode: string;
    hideArchitectureSearchFlg: Boolean;
    architectureSearchType: string;
    architectureSearchLabel: string;
    allowSecurityMarkingsFlg: Boolean;
    securityBannerText: string;
    securityBannerBackgroundColor: string;
    securityBannerTextColor: string;
    userInputWarning: string;
    submissionFormWarning: string;
    changeRequestWarning: string;
    feedbackHandler: string;
    primaryColor: string;
    primaryTextColor: string;
    secondaryColor: string;
    accentColor: string;
    quoteColor: string;
    linkColor: string;
    linkVisitedColor: string;
    linkhoverColor: string;
    panelHeaderColor: string;
    panelHeaderTextColor: string;
    overrideCSS: string;
    useDefaultLandingPage: Boolean;
    showSupportMenuOnLogin: Boolean;
    landingTemplate: string; //LandingTemplate
}